Rename Poly1305 => RbNaCl::OneTimeAuths::Poly1305
This follows the naming convention from the C source code.

"OneTimeAuths" seems a bit silly but it is more consistent. Plus it's not a name
that end users will have to deal with directly (at least in a typical case)
